Revert "Remove references to /python/table, move perspective-python package closer to top level, fix python scripts"
This reverts commit 78a9e355da885a2570d12dd45da91dd85e724409.
