+++ package.json
@@ -59,12 +59,15 @@
         "build": "node scripts/build.js",
         "build_cpp": "node scripts/build_cpp.js",
         "build_python": "node scripts/build_python.js",
-        "install_python": "cd python/ && python3 setup.py install",
+        "build_python:table": "node scripts/build_python_table.js",
+        "install_python": "cd python/table && python3 setup.py install",
+        "install_python:table": "cd python/perspective && python3 setup.py install",
         "bench": "node scripts/bench.js",
         "docs": "node scripts/docs.js",
         "test": "node scripts/test.js",
         "test_cpp": "node scripts/test_cpp.js",
         "test_python": "node scripts/test_python.js",
+        "test_python:table": "node scripts/test_python_table.js",
         "clean": "node scripts/clean.js",
         "start": "lerna run start --stream --scope",
         "precommit": "npm run lint",
@@ -73,6 +76,7 @@
         "lint:eslint": "eslint packages/*/src/**/*.js packages/*/test/**/*.js examples/*/*.js",
         "lint:cpp": "node scripts/lint_cpp.js",
         "lint_python": "node scripts/lint_python.js",
+        "lint_python:table": "node scripts/lint_python_table.js",
         "fix:es": "npm run lint:eslint -- --fix",
         "fix:cpp": "node scripts/fix_cpp.js",
         "fix:less": "prettier --tab-width 4 --write packages/**/src/less/*.less",
