+++ python/table/setup.cfg
