+++ scripts/lint_python.js
@@ -17,14 +17,13 @@ function docker(image = "emsdk") {
     if (process.env.PSP_CPU_COUNT) {
         cmd += ` --cpus="${parseInt(process.env.PSP_CPU_COUNT)}.0"`;
     }
-    cmd += ` -v $(pwd):/usr/src/app/python/ -w /usr/src/app/python/ perspective/${image}`;
+    cmd += ` -v $(pwd):/usr/src/app/python/perspective -w /usr/src/app/python/perspective perspective/${image}`;
     return cmd;
 }
 
 try {
-    let cmd = "cd python/ && python3 -m flake8 perspective && echo OK";
+    let cmd = "cd python/perspective && python3 -m flake8 perspective && echo OK";
     if (process.env.PSP_DOCKER) {
-        // eslint-disable-next-line prettier/prettier
         execute(docker("python") + " bash -c \"" + cmd + '\"');
     } else {
         execute(cmd);
